+++ src/i8c/constants.py
@@ -38,7 +38,6 @@ I8_TYPE_PTR = "p"
 I8_TYPE_OPAQUE = "o"
 I8_TYPE_FUNC = "F"
 
-I8_EXT_RELADDR = "x" # XXX deprecated
 I8_EXT_FUNCTION = "f"
 I8_EXT_SYMBOL = "s"